requirements.txt
sklearn
pandas
azureml-sdk